Fix problem with recent commit for JERI mux fix for jtreg stability tests, In this case an int was cast to a byte (this was not intended) and then compared to an int. This cast has been removed, it was an unintended coding error. This has been fixed because the byte would overflow and prevent the comparison from working, it would have been a critical bug.  TODO: for River 3.0.1 create a test case to ensure this section of code is tested, no tests presently confirm whether the condition in this section of code is met.
git-svn-id: https://svn.apache.org/repos/asf/river/jtsk/trunk@1719239 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/river/jeri/internal/mux/MuxClient.java
@@ -69,7 +69,7 @@ public class MuxClient extends Mux {
 		ioe.initCause(muxDownCause);
 		throw ioe;
 	    }
-	    byte sessionID = (byte) busySessions.nextClearBit(0);
+	    int sessionID = busySessions.nextClearBit(0);
 	    if (sessionID > Mux.MAX_SESSION_ID) {
 		throw new IOException("no free sessions");
 	    }
